Fixed intruding aircraft symbol drawing. Did some more refactoring/code clean up.

+++ RecommendationRange.h
@@ -1,7 +1,7 @@
 #pragma once
 
 typedef struct RecommendationRange {
-	float min_vertical_speed;
-	float max_vertical_speed;
+	double min_vertical_speed;
+	double max_vertical_speed;
 	bool recommended;
 } RecommendationRange;
